requirements.txt
openai
python-dotenv
discord